bugfix(volo-http): do not use sync resolver from `hickory_resolver`
The `hickory_resolver::Resolver` is a sync resolver, and
`Resolver::resolve` is a sync function, but it uses
`Runtime::block_on` of `tokio::runtime::Runtime`, which will **panic**
when it is in the scope of a runtime.

This commit implements a `block_on` function, which will check if
there is any runtime. If there is a runtime, just use its `block_on`,
otherwise create a new one and call `block_on`. This function can
ensure that the DNS resolver will never panic on runtime.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>